+++ frontend/src/components/ui/EngineSelect.tsx
@@ -2,7 +2,7 @@
 
 import clsx from 'clsx';
 import { createPortal } from 'react-dom';
-import type { KeyboardEvent as ReactKeyboardEvent, MouseEvent as ReactMouseEvent, ReactNode } from 'react';
+import type { KeyboardEvent as ReactKeyboardEvent } from 'react';
 import {
   useCallback,
   useEffect,
@@ -12,230 +12,25 @@ import {
   useState,
   useId
 } from 'react';
-import type { EngineAvailability, EngineCaps, Mode, Resolution } from '@/types/engines';
+import type { EngineAvailability, EngineCaps, Mode } from '@/types/engines';
 import { Card } from './Card';
 import { Chip } from './Chip';
 import { EngineIcon } from '@/components/ui/EngineIcon';
-import type { FalEngineEntry } from '@/config/falEngines';
 import { useI18n } from '@/lib/i18n/I18nProvider';
-import { Link } from '@/i18n/navigation';
-import { normalizeEngineId } from '@/lib/engine-alias';
-import { DEFAULT_ENGINE_GUIDE } from '@/lib/engine-guides';
-import { getEngineSelectFamilyRank } from '@/lib/engine-family-priority';
-import { getExamplesHref } from '@/lib/examples-links';
-import { formatResolutionList } from '@/lib/resolution-labels';
-import { getLocalizedModeLabel, normalizeUiLocale } from '@/lib/ltx-localization';
-
-type EngineRegistryMeta = {
-  order: Map<string, number>;
-  meta: Map<string, FalEngineEntry>;
-};
-
-let engineRegistryMetaCache: EngineRegistryMeta | null = null;
-let engineRegistryMetaPromise: Promise<EngineRegistryMeta> | null = null;
-
-async function ensureEngineRegistryMeta(): Promise<EngineRegistryMeta> {
-  if (engineRegistryMetaCache) return engineRegistryMetaCache;
-  if (!engineRegistryMetaPromise) {
-    engineRegistryMetaPromise = import('@/config/falEngines')
-      .then((mod) => {
-        const registry = mod.listFalEngines();
-        const meta: EngineRegistryMeta = {
-          order: new Map(registry.map((entry, index) => [entry.id, index])),
-          meta: new Map(registry.map((entry) => [entry.id, entry])),
-        };
-        engineRegistryMetaCache = meta;
-        return meta;
-      })
-      .catch((error) => {
-        engineRegistryMetaPromise = null;
-        throw error;
-      });
-  }
-  return engineRegistryMetaPromise;
-}
-
-const ENGINE_VARIANT_LABEL_OVERRIDES: Record<string, string> = {
-  'kling-3-standard': 'Standard',
-  'kling-3-pro': 'Pro',
-  'ltx-2-3': 'Pro',
-  'ltx-2-3-fast': 'Fast',
-};
-const ENGINE_VARIANT_SORT_ORDER = new Map<string, number>([
-  ['ltx-2-3', 0],
-  ['ltx-2-3-fast', 1],
-]);
-const ENGINE_LEGACY_STORAGE_KEY = 'engineSelect.showLegacy';
-
-const DEFAULT_MODE_OPTIONS: Mode[] = ['t2v', 'i2v', 'v2v', 'reframe', 'ref2v', 'fl2v', 'extend', 'a2v', 'retake', 'r2v'];
-
-const ENGINE_MODE_LABEL_OVERRIDES: Record<string, Partial<Record<Mode, string>>> = {
-  lumaRay2: {
-    v2v: 'Modify',
-    reframe: 'Reframe',
-  },
-  lumaRay2_flash: {
-    v2v: 'Modify',
-    reframe: 'Reframe',
-  },
-  'veo-3-1': {
-    ref2v: 'Reference',
-    fl2v: 'First/Last',
-    extend: 'Extend',
-  },
-  'veo-3-1-fast': {
-    fl2v: 'First/Last',
-    extend: 'Extend',
-  },
-  'veo-3-1-lite': {
-    fl2v: 'First/Last',
-  },
-  'kling-2-5-turbo': {
-    t2v: 'Pro · Text',
-    i2v: 'Pro · Image',
-    i2i: 'Standard · Image',
-  },
-  'wan-2-5': {
-    t2v: 'Text · Audio-ready',
-    i2v: 'Image · Audio-ready',
-  },
-  'wan-2-6': {
-    t2v: 'Text · Multi-shot',
-    i2v: 'Image · Animate',
-    r2v: 'Reference · Consistency',
-  },
-};
-
-function getModeLabel(
-  engineId: string | undefined,
-  value: Mode,
-  locale: string | undefined,
-  overrides?: Partial<Record<Mode, string>>
-): string {
-  const custom = overrides?.[value];
-  if (custom) return custom;
-  const engineOverrides = engineId ? ENGINE_MODE_LABEL_OVERRIDES[engineId] : undefined;
-  return engineOverrides?.[value] ?? getLocalizedModeLabel(value, normalizeUiLocale(locale));
-}
-
-function getModeDisplayOrder(engineId: string | undefined, modes: Mode[]): Mode[] {
-  if (engineId === 'lumaRay2' || engineId === 'lumaRay2_flash') {
-    const order: Mode[] = ['t2v', 'i2v', 'v2v', 'reframe'];
-    return order.filter((mode) => modes.includes(mode));
-  }
-  if (engineId === 'veo-3-1') {
-    const order: Mode[] = ['t2v', 'i2v', 'ref2v', 'fl2v', 'extend'];
-    return order.filter((mode) => modes.includes(mode));
-  }
-  if (engineId === 'veo-3-1-fast') {
-    const order: Mode[] = ['t2v', 'i2v', 'fl2v', 'extend'];
-    return order.filter((mode) => modes.includes(mode));
-  }
-  if (engineId === 'veo-3-1-lite') {
-    const order: Mode[] = ['t2v', 'i2v', 'fl2v'];
-    return order.filter((mode) => modes.includes(mode));
-  }
-  return modes;
-}
-
-function formatAvgDuration(value: number | null | undefined): string | null {
-  if (typeof value !== 'number' || !Number.isFinite(value) || value <= 0) return null;
-  const seconds = value / 1000;
-  const precision = seconds < 10 ? 1 : 0;
-  return `${seconds.toFixed(precision)}s`;
-}
-
-function getEngineDiscoveryRank(engineId: string, registryMeta: EngineRegistryMeta | null): number {
-  return registryMeta?.meta.get(engineId)?.surfaces.app.discoveryRank ?? Number.MAX_SAFE_INTEGER;
-}
-
-function compareEnginesByDefaultPriority(
-  a: EngineCaps,
-  b: EngineCaps,
-  registryMeta: EngineRegistryMeta | null,
-  engineMeta?: Map<string, FalEngineEntry>
-): number {
-  const metaA = engineMeta?.get(a.id) ?? registryMeta?.meta.get(a.id);
-  const metaB = engineMeta?.get(b.id) ?? registryMeta?.meta.get(b.id);
-  const familyRankA = getEngineSelectFamilyRank(metaA);
-  const familyRankB = getEngineSelectFamilyRank(metaB);
-  if (familyRankA !== familyRankB) {
-    if (familyRankA === Number.MAX_SAFE_INTEGER) return 1;
-    if (familyRankB === Number.MAX_SAFE_INTEGER) return -1;
-    return familyRankA - familyRankB;
-  }
-  const variantOrderA = ENGINE_VARIANT_SORT_ORDER.get(a.id);
-  const variantOrderB = ENGINE_VARIANT_SORT_ORDER.get(b.id);
-  if (variantOrderA != null || variantOrderB != null) {
-    if (variantOrderA == null) return 1;
-    if (variantOrderB == null) return -1;
-    if (variantOrderA !== variantOrderB) return variantOrderA - variantOrderB;
-  }
-  const priorityOrderA = getEngineDiscoveryRank(a.id, registryMeta);
-  const priorityOrderB = getEngineDiscoveryRank(b.id, registryMeta);
-  if (priorityOrderA !== priorityOrderB) {
-    if (priorityOrderA === Number.MAX_SAFE_INTEGER) return 1;
-    if (priorityOrderB === Number.MAX_SAFE_INTEGER) return -1;
-    return priorityOrderA - priorityOrderB;
-  }
-  const orderA = registryMeta?.order.get(a.id) ?? Number.MAX_SAFE_INTEGER;
-  const orderB = registryMeta?.order.get(b.id) ?? Number.MAX_SAFE_INTEGER;
-  return orderA - orderB;
-}
-
-interface EngineSelectProps {
-  engines: EngineCaps[];
-  engineId: string;
-  onEngineChange: (engineId: string) => void;
-  mode: Mode;
-  onModeChange: (mode: Mode) => void;
-  modeOptions?: Mode[];
-  showBillingNote?: boolean;
-  modeLabelOverrides?: Partial<Record<Mode, string>>;
-  showModeSelect?: boolean;
-  modeLayout?: 'inline' | 'stacked';
-  variant?: 'card' | 'bar';
-  density?: 'default' | 'compact';
-  className?: string;
-}
-
-const DEFAULT_ENGINE_SELECT_COPY = {
-  avgDuration: 'Avg {value}',
-  choose: 'Choose engine',
-  variant: 'Variant',
-  browse: 'Browse engines...',
-  inputMode: 'Input mode',
-  unsupportedMode: 'Not supported by this engine',
-  modal: {
-    close: 'Close',
-    title: 'Choose the right engine for your shot',
-    subtitle:
-      'Each model has its own strengths - some are fast, others cinematic or experimental. See what fits your project, then generate with confidence.',
-    pricingLink: 'How pricing works',
-    searchPlaceholder: 'Search by engine, provider, or capability',
-    modeAll: 'Mode: All',
-    modeValue: 'Mode: {value}',
-    resolutionAll: 'Resolution: All',
-    legacyToggleLabel: 'Legacy models',
-    viewModel: 'View model page',
-    viewExamples: 'View examples',
-    empty:
-      'No engines match your filters yet. Adjust the filters or clear the search to explore the full catalogue.',
-    disclaimer: 'Logos are used for descriptive purposes only. Trademarks belong to their respective owners.',
-    descriptionFallback:
-      'Versatile engine ready for price-before-you-generate workflows. Review specs and run with confidence.',
-  },
-  guides: DEFAULT_ENGINE_GUIDE,
-} as const;
-
-type EngineSelectCopy = typeof DEFAULT_ENGINE_SELECT_COPY;
-
-
-interface DropdownPosition {
-  top: number;
-  left: number;
-  width: number;
-}
+import { BrowseEnginesModal } from './engine-select/BrowseEnginesModal';
+import { DEFAULT_ENGINE_SELECT_COPY, type EngineSelectCopy } from './engine-select/engine-select-copy';
+import {
+  compareEnginesByDefaultPriority,
+  DEFAULT_MODE_OPTIONS,
+  ENGINE_LEGACY_STORAGE_KEY,
+  ENGINE_VARIANT_LABEL_OVERRIDES,
+  ensureEngineRegistryMeta,
+  formatAvgDuration,
+  getCachedEngineRegistryMeta,
+  getModeDisplayOrder,
+  getModeLabel,
+} from './engine-select/engine-select-helpers';
+import type { DropdownPosition, EngineRegistryMeta, EngineSelectProps } from './engine-select/engine-select-types';
 
 export function EngineSelect({
   engines,
@@ -254,7 +49,7 @@ export function EngineSelect({
 }: EngineSelectProps) {
   const { t, locale } = useI18n();
   const copy = t('workspace.generate.engineSelect', DEFAULT_ENGINE_SELECT_COPY) as EngineSelectCopy;
-  const [registryMeta, setRegistryMeta] = useState<EngineRegistryMeta | null>(() => engineRegistryMetaCache);
+  const [registryMeta, setRegistryMeta] = useState<EngineRegistryMeta | null>(() => getCachedEngineRegistryMeta());
   const [open, setOpen] = useState(false);
   const [position, setPosition] = useState<DropdownPosition | null>(null);
   const [highlightedIndex, setHighlightedIndex] = useState<number>(-1);
@@ -905,384 +700,5 @@ export function EngineSelect({
     <Card ref={containerRef} className={containerClassName}>
       {content}
     </Card>
-  );
-}
-
-interface BrowseEnginesModalProps {
-  engines: EngineCaps[];
-  selectedEngineId: string;
-  onClose: () => void;
-  onSelect: (engineId: string) => void;
-  copy: EngineSelectCopy;
-  modeLabelOverrides?: Partial<Record<Mode, string>>;
-  engineMeta?: Map<string, FalEngineEntry>;
-  showLegacy: boolean;
-  onToggleLegacy: () => void;
-}
-
-type ModeFilter = 'all' | Mode;
-
-function BrowseEnginesModal({
-  engines,
-  selectedEngineId,
-  onClose,
-  onSelect,
-  copy,
-  modeLabelOverrides,
-  engineMeta,
-  showLegacy,
-  onToggleLegacy,
-}: BrowseEnginesModalProps) {
-  const { locale } = useI18n();
-  const modalCopy = copy.modal;
-  const viewModelLabel = modalCopy.viewModel ?? DEFAULT_ENGINE_SELECT_COPY.modal.viewModel;
-  const viewExamplesLabel = modalCopy.viewExamples ?? DEFAULT_ENGINE_SELECT_COPY.modal.viewExamples;
-  const legacyToggleLabel =
-    modalCopy.legacyToggleLabel ?? DEFAULT_ENGINE_SELECT_COPY.modal.legacyToggleLabel;
-  const [portalElement, setPortalElement] = useState<HTMLDivElement | null>(null);
-  const [searchTerm, setSearchTerm] = useState('');
-  const [modeFilter, setModeFilter] = useState<ModeFilter>('all');
-  const [resolutionFilter, setResolutionFilter] = useState<'all' | Resolution>('all');
-
-  useEffect(() => {
-    const element = document.createElement('div');
-    element.dataset.engineBrowsePortal = 'true';
-    document.body.appendChild(element);
-    setPortalElement(element);
-    return () => {
-      if (element.parentNode) {
-        element.parentNode.removeChild(element);
-      }
-    };
-  }, []);
-
-  useEffect(() => {
-    const previous = document.body.style.overflow;
-    document.body.style.overflow = 'hidden';
-    return () => {
-      document.body.style.overflow = previous;
-    };
-  }, []);
-
-  const legacyFilteredEngines = useMemo(() => {
-    return engines.filter((engine) => {
-      const meta = engineMeta?.get(engine.id);
-      if (!meta?.isLegacy) return true;
-      if (showLegacy) return true;
-      return engine.id === selectedEngineId;
-    });
-  }, [engines, engineMeta, selectedEngineId, showLegacy]);
-
-  useEffect(() => {
-    const handleKey = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        onClose();
-      }
-    };
-    document.addEventListener('keydown', handleKey);
-    return () => document.removeEventListener('keydown', handleKey);
-  }, [onClose]);
-
-  const resolutions = useMemo<Resolution[]>(() => {
-    const set = new Set<Resolution>();
-    legacyFilteredEngines.forEach((engine) => {
-      engine.resolutions?.forEach((resolution) => set.add(resolution));
-    });
-
-    const resolutionRank = (value: Resolution) => {
-      const raw = value.toString().toLowerCase();
-      const explicit: Record<string, number> = {
-        auto: -1,
-        portrait_hd: 720,
-        landscape_hd: 720,
-        square_hd: 720,
-      };
-
-      if (raw in explicit) return explicit[raw];
-      if (raw.endsWith('k')) {
-        const numeric = Number(raw.replace('k', ''));
-        return Number.isFinite(numeric) ? numeric * 1000 : 0;
-      }
-
-      const match = raw.match(/(\d+)\s*p/);
-      if (match) return Number(match[1]);
-      const fallback = Number(raw.replace(/[^\d.]/g, ''));
-      return Number.isFinite(fallback) ? fallback : 0;
-    };
-
-    return Array.from(set).sort((a, b) => {
-      const aRank = resolutionRank(a);
-      const bRank = resolutionRank(b);
-      if (aRank !== bRank) return bRank - aRank;
-      return a.localeCompare(b);
-    });
-  }, [legacyFilteredEngines]);
-
-  const searchValue = searchTerm.trim().toLowerCase();
-
-  const filteredEngines = useMemo(() => {
-    const guides = copy.guides ?? DEFAULT_ENGINE_GUIDE;
-    const ranked = legacyFilteredEngines
-      .slice()
-      .filter((engine) => {
-        if (modeFilter !== 'all' && !engine.modes.includes(modeFilter)) return false;
-        if (resolutionFilter !== 'all' && !engine.resolutions.includes(resolutionFilter)) return false;
-
-        const meta = engineMeta?.get(engine.id);
-        if (!searchValue) return true;
-        const guide = guides[engine.id];
-        const haystack = [
-          meta?.marketingName ?? engine.label ?? engine.id,
-          engine.provider,
-          engine.version ?? '',
-          guide?.description ?? '',
-          ...(guide?.badges ?? []),
-          meta?.seo.description ?? '',
-        ]
-          .join(' ')
-          .toLowerCase();
-        return haystack.includes(searchValue);
-      })
-      .sort((a, b) => {
-        const ranked = compareEnginesByDefaultPriority(a, b, null, engineMeta);
-        if (ranked !== 0) return ranked;
-        if (a.isLab !== b.isLab) return a.isLab ? 1 : -1;
-        const aName = engineMeta?.get(a.id)?.marketingName ?? a.label ?? a.id;
-        const bName = engineMeta?.get(b.id)?.marketingName ?? b.label ?? b.id;
-        return aName.localeCompare(bName);
-      });
-    return ranked;
-  }, [legacyFilteredEngines, modeFilter, resolutionFilter, searchValue, copy, engineMeta]);
-
-  const handleBackdropClick = useCallback(
-    (event: ReactMouseEvent<HTMLDivElement>) => {
-      if (event.target === event.currentTarget) {
-        onClose();
-      }
-    },
-    [onClose]
-  );
-
-
-  const FilterChip = ({
-    active,
-    onClick,
-    children,
-  }: {
-    active: boolean;
-    onClick: () => void;
-    children: ReactNode;
-  }) => (
-    <button
-      type="button"
-      onClick={onClick}
-      className={clsx(
-        'rounded-input border px-3 py-1.5 text-xs font-semibold transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring',
-        active
-          ? 'border-brand bg-brand text-on-brand'
-          : 'border-border bg-surface text-text-secondary hover:border-border-hover hover:bg-surface-2 hover:text-text-primary'
-      )}
-    >
-      {children}
-    </button>
-  );
-
-  if (!portalElement) {
-    return null;
-  }
-
-  return createPortal(
-    <div
-      className="fixed inset-0 z-[10000] flex items-center justify-center bg-black/50 px-4 py-6"
-      role="dialog"
-      aria-modal="true"
-      onClick={handleBackdropClick}
-    >
-      <div className="relative flex max-h-[95vh] w-full max-w-6xl flex-col overflow-y-auto rounded-modal border border-border bg-surface shadow-float">
-        <button
-          type="button"
-          onClick={onClose}
-          className="absolute right-4 top-4 rounded-full border border-hairline bg-surface px-3 py-1 text-xs font-medium text-text-secondary transition hover:bg-surface-2 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-        >
-          {modalCopy.close}
-        </button>
-        <header className="border-b border-hairline bg-bg px-6 pb-5 pt-6">
-          <div className="flex flex-wrap items-start justify-between gap-4">
-            <div className="max-w-xl space-y-2">
-              <h2 className="text-xl font-semibold text-text-primary">{modalCopy.title}</h2>
-              <p className="text-sm text-text-secondary">{modalCopy.subtitle}</p>
-            </div>
-            <div className="flex flex-col items-end gap-2">
-              <a
-                href="/docs/pricing"
-                target="_blank"
-                rel="noreferrer"
-                className="text-xs font-semibold text-brand hover:underline"
-              >
-                {modalCopy.pricingLink}
-              </a>
-            </div>
-          </div>
-          <div className="mt-6 stack-gap-sm">
-            <div className="relative">
-              <input
-                type="search"
-                value={searchTerm}
-                onChange={(event) => setSearchTerm(event.target.value)}
-                placeholder={modalCopy.searchPlaceholder}
-                className="w-full rounded-input border border-border bg-surface px-3 py-3 text-sm text-text-primary shadow-sm focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-              />
-            </div>
-            <div className="flex flex-wrap gap-2">
-              <FilterChip active={modeFilter === 'all'} onClick={() => setModeFilter('all')}>
-                {modalCopy.modeAll}
-              </FilterChip>
-              {DEFAULT_MODE_OPTIONS.map((candidate) => (
-                <FilterChip key={candidate} active={modeFilter === candidate} onClick={() => setModeFilter(candidate)}>
-                  {modalCopy.modeValue.replace('{value}', candidate.toUpperCase())}
-                </FilterChip>
-              ))}
-            </div>
-            <div className="flex flex-wrap gap-2">
-              <FilterChip active={resolutionFilter === 'all'} onClick={() => setResolutionFilter('all')}>
-                {modalCopy.resolutionAll}
-              </FilterChip>
-              {resolutions.map((resolution) => (
-                <FilterChip
-                  key={resolution}
-                  active={resolutionFilter === resolution}
-                  onClick={() => setResolutionFilter(resolution)}
-                >
-                  {resolution}
-                </FilterChip>
-              ))}
-            </div>
-            <div className="flex flex-wrap gap-2">
-              <FilterChip active={showLegacy} onClick={onToggleLegacy}>
-                {legacyToggleLabel}
-              </FilterChip>
-            </div>
-          </div>
-        </header>
-        <div className="bg-surface px-6 py-6">
-          <div className="grid grid-gap-sm md:grid-cols-2">
-            {filteredEngines.map((engine) => {
-              const guide = (copy.guides ?? DEFAULT_ENGINE_GUIDE)[engine.id];
-              const isSelected = engine.id === selectedEngineId;
-              const badges = guide?.badges ?? [];
-              const labsBadgeNeeded = engine.isLab && !badges.some((badge) => badge === 'Labs');
-              const combinedBadges = labsBadgeNeeded ? [...badges, 'Labs'] : badges;
-              const canonicalId = normalizeEngineId(engine.id) ?? engine.id;
-              const meta = engineMeta?.get(canonicalId);
-              const name = meta?.marketingName ?? engine.label ?? engine.id;
-              const versionLabel = meta?.versionLabel ?? engine.version ?? '-';
-              const description = guide?.description ?? meta?.seo.description ?? modalCopy.descriptionFallback;
-              const avgDurationLabel = formatAvgDuration(engine.avgDurationMs);
-              const modelSlug = meta?.modelSlug;
-              const modelHref = modelSlug ? { pathname: '/models/[slug]', params: { slug: modelSlug } } : null;
-              const allowExamples = meta?.category !== 'image' && meta?.type !== 'image';
-              const examplesHref = modelSlug && allowExamples ? getExamplesHref(modelSlug) : null;
-              const showModelLink = Boolean(modelHref);
-              const showExamplesLink = Boolean(examplesHref);
-
-              return (
-                <Card
-                  key={engine.id}
-                  className={clsx(
-                    'flex cursor-pointer flex-col gap-4 overflow-hidden p-5 transition hover:border-text-muted hover:bg-surface-2 hover:shadow-float',
-                    isSelected && 'border-brand bg-surface-2 shadow-float'
-                  )}
-                  onClick={() => onSelect(engine.id)}
-                >
-                  <div className="flex flex-1 flex-col gap-4">
-                    <div className="flex items-start gap-4">
-                      <EngineIcon engine={engine} size={44} className="shrink-0" />
-                      <div className="flex flex-1 items-start justify-between gap-4">
-                        <div className="space-y-1">
-                          <h3 className="text-base font-semibold text-text-primary">{name}</h3>
-                          <p className="text-xs uppercase tracking-micro text-text-muted">
-                            {engine.provider} - {versionLabel}
-                          </p>
-                        </div>
-                        <div className="flex flex-col items-end gap-1 text-[11px] text-text-muted">
-                          {avgDurationLabel && (
-                            <span className="rounded-input border border-border px-2 py-0.5">
-                              {copy.avgDuration.replace('{value}', avgDurationLabel)}
-                            </span>
-                          )}
-                          {engine.status && (
-                            <span className="rounded-input border border-border px-2 py-0.5 uppercase tracking-micro">
-                              {engine.status}
-                            </span>
-                          )}
-                        </div>
-                      </div>
-                    </div>
-                    <p className="text-sm leading-6 text-text-secondary">{description}</p>
-                    <div className="flex flex-wrap gap-2">
-                      {combinedBadges.map((badge) => (
-                        <span
-                          key={badge}
-                          className="inline-flex items-center rounded-full bg-bg px-3 py-1 text-[12px] font-medium text-text-secondary"
-                        >
-                          {badge}
-                        </span>
-                      ))}
-                    </div>
-                    <div className="flex flex-wrap items-center gap-4 text-xs text-text-muted">
-                      <span>
-                        Modes:{' '}
-                        {getModeDisplayOrder(engine.id, engine.modes)
-                          .map((entry) => getModeLabel(engine.id, entry, locale, modeLabelOverrides))
-                          .join(' / ')}
-                      </span>
-                      <span>
-                        Max {engine.maxDurationSec}s / Res {formatResolutionList(engine.id, engine.resolutions).join(' / ')}
-                      </span>
-                    </div>
-                  </div>
-                  {(showModelLink || showExamplesLink) && (
-                    <div className="mt-auto -mx-5 -mb-5 border-t border-hairline">
-                      <div className="flex text-xs font-semibold">
-                        {modelHref && (
-                          <Link
-                            href={modelHref}
-                            className={clsx(
-                              'flex-1 px-4 py-2 text-center text-text-secondary transition hover:text-text-primary',
-                              showExamplesLink ? 'border-r border-hairline' : null,
-                              'bg-surface',
-                              'hover:bg-surface-2'
-                            )}
-                            onClick={(event) => event.stopPropagation()}
-                          >
-                            {viewModelLabel}
-                          </Link>
-                        )}
-                        {examplesHref && (
-                          <Link
-                            href={examplesHref}
-                            className="flex-1 bg-surface px-4 py-2 text-center text-text-secondary transition hover:bg-surface-2 hover:text-text-primary"
-                            onClick={(event) => event.stopPropagation()}
-                          >
-                            {viewExamplesLabel}
-                          </Link>
-                        )}
-                      </div>
-                    </div>
-                  )}
-                </Card>
-              );
-            })}
-            {!filteredEngines.length && (
-              <div className="col-span-full rounded-input border border-dashed border-border bg-bg px-6 py-12 text-center text-sm text-text-muted">
-                {modalCopy.empty}
-              </div>
-            )}
-          </div>
-          <p className="mt-6 text-center text-[11px] text-text-muted">{modalCopy.disclaimer}</p>
-        </div>
-      </div>
-    </div>,
-    portalElement
   );
 }
